+++ api/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2019-11-07
+
+### Added
+
+- [DEV-8049] Added `wise_units-ffi` crate to expose C bindings for wise_units functions, 
+  with C-safe wrapper types around `Measurement` and `Unit`.
+- Derive `Clone` for `error::Error` and `parser::Error`.
+
 ## [0.11.0] - 2019-07-17
 
 ### Added